* added string_hash.hpp (outsourced from SysteMoC)
git-archimport-id: falk@cs.fau.de--2007-common-intern/HscdSupport--mainline--0.6--patch-100

+++ pkginclude/cosupport/SMXIdManager.hpp
@@ -10,7 +10,7 @@
 namespace CoSupport {
 
 /// @brief Identifier for elements
-typedef uint64_t SMXId;
+typedef uint32_t SMXId;
 
 /**
  * @brief Helper struct for (de)serializing Ids
